requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
cffi==1.17.1
colorama==0.4.6
cryptography==45.0.2
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
MouseInfo==0.1.3
mss==10.0.0
numpy==2.2.6
openai==1.79.0
opencv-python==4.11.0.86
packaging==25.0
pefile==2023.2.7
pillow==11.2.1
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyGetWindow==0.0.9
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyMsgBox==1.0.9
pynput==1.8.1
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.1.0
pytweening==1.2.0
pywin32-ctypes==0.2.3
qasync==0.27.1
qrcode==8.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
